test: split the OpenRouter reasoning tests out and widen the parity scan
zai/tests.rs was already 43 lines over its file-size ceiling at HEAD; the
new effort tests would have pushed it further, so both they and the
existing wire-body test move into zai/tests/openrouter_effort_tests.rs —
one subject per module, and the file lands genuinely under its ceiling
(1781 -> 1685) instead of needing a raised baseline.

That move exposed a blind spot in provider_parity's witness scan: it reads
source TEXT from a fixed list that never included the zai/tests/
submodules, so a witness that stays a real passing test but moves into a
split-out module reads as a rotted proof. All three submodules are now
listed.

Also unbreaks stella-tui's theme test, which referenced palette::EMBER and
EMBER_DEEP — names retired when the palette moved to the brand blue, so it
had not compiled since. LIGHT_REMAP is the authority on the value.
